Refactor models into arch specific APIs
