+++ lm_eval/models/openai_completion.py
@@ -52,7 +52,7 @@ def oa_completion(**kwargs):
             backoff_time *= 1.5
 
 
-class GPT3LM(BaseLM):
+class CompletionLM(BaseLM):
     REQ_CHUNK_SIZE = 20
 
     def __init__(self, engine, device=None, truncate=False, parallelize=False):
